Removed `InheritedModelAspectListener`
